Squashed commit of the following:
commit 1e6313ff2271ac9dfed9f64f467669e2feb5ccb1
Author: Rumesh Udash <allmesh67@gmail.com>
Date:   Mon Aug 19 13:09:55 2019 +0545

    Version 1.6.3 and Grunt Task

commit a7e4adede6b3d8d23c722dc606a80fc9a94e2167
Author: Rumesh Udash <allmesh67@gmail.com>
Date:   Mon Aug 19 12:46:46 2019 +0545

    Tweak - Template file update to locate add-ons file (#164)

commit 3d68fef10817179deb47c6ee08462454df2e1665
Author: Lekhnath Pandey <46083608+lekhnath005@users.noreply.github.com>
Date:   Mon Aug 19 12:41:27 2019 +0545

    Fix - Email format issue (#163)

commit a4f2dacbd008163da3d18b67162c3d8ae9b30233
Author: Lekhnath Pandey <46083608+lekhnath005@users.noreply.github.com>
Date:   Mon Aug 19 12:38:56 2019 +0545

    enhancement/profile changed email (#158)

    * Enhancement/Profile changed email

    * code cleaning

    * code cleaning

    * Fix - enable profile detail change option key

    * Cleanup

    * bug fixing - send emails to multiple admins

    * fix email format

    * fix send email to multiple admin

commit 0c746bcbb2d131fb2d1afe991ac6058fc9fa9553
Author: Rumesh Udash <allmesh67@gmail.com>
Date:   Mon Aug 19 11:05:54 2019 +0545

    Feature - Fullscreen builder mode (#154)

    * Feature - Fullscreen builder mode

    * Esc key event for closing fullscreen mode

    * Tweak - UI enhancements for full screen mode

    * Tweak - Input field box design on builder

    * Fix - Notice position issue

commit 873440ad292300595d89c10eb1a69aa2ef80b2c3
Author: Rumesh Udash <allmesh67@gmail.com>
Date:   Fri Aug 16 16:01:09 2019 +0545

    Add - Accept redirect_to get parameter in login page (#161)

commit 2327e74062c13e2d2dab5f1eaf8aa548d504c109
Author: Rumesh Udash <allmesh67@gmail.com>
Date:   Fri Aug 16 15:59:10 2019 +0545

    Profile details field hook (#162)

    * Add - Accept redirect_to get parameter in login page

    * Add hook to filter form data in profile details

    * cleanup

commit 7fa9de48766b69a0a513ce0485c00a2a30069597
Author: Lekhnath Pandey <46083608+lekhnath005@users.noreply.github.com>
Date:   Fri Aug 16 15:44:40 2019 +0545

    Fix/pw strength error msg (#157)

    * password strength error message issue fixed

    * fix bug and code cleaning

    * focus out bug fix

    * bug fix - Password strenght message changed

    * Cleanup

commit fca196bf5623dfe424caf57b36d67404780ed493
Author: Nilan Bhasima <nilanbhas@gmail.com>
Date:   Fri Aug 2 15:10:38 2019 +0545

    Line height for input field

commit e41a32f0be26cf6b776930c17cf69a306e5a9ca0
Author: Rumesh Udash <allmesh67@gmail.com>
Date:   Fri Aug 2 11:01:04 2019 +0545

    Fix - Profile picture issue in dashboard

commit 88e3caca8c4b1949f1ebbda3f89f08d02bccda17
Author: Rumesh Udash <allmesh67@gmail.com>
Date:   Fri Aug 2 09:39:34 2019 +0545

    Form end hook (#159)

    * Form start and end hook

    * Escaping form id

commit 331f759df4935cd33868d7c6766d38819ea3cbf2
Author: Rumesh Udash <allmesh67@gmail.com>
Date:   Wed Jul 31 12:12:37 2019 +0545

    Fix - Profile picture save without selecting any file

commit 0af8d3642371d4199bb0da0d46023084c67d6fbc
Author: Rabin Shrestha <sun.ravi90@gmail.com>
Date:   Thu Jul 25 17:20:11 2019 +0545

    Update package.json

commit aa0fd3a3c0fb2b075fef3b9931fabfa209d3201b
Merge: 731a0875 7dc1bef6
Author: Rabin Shrestha <sun.ravi90@gmail.com>
Date:   Thu Jul 25 17:10:20 2019 +0545

    Merge branch 'develop' of github.com:wpeverest/user-registration into develop

commit 731a0875ff5d6aef727a232aca1c3bc99f0b0b16
Author: Rabin Shrestha <sun.ravi90@gmail.com>
Date:   Thu Jul 25 17:10:10 2019 +0545

    Update plugin description

commit 7dc1bef6564d8bda143eea3321f101e0e1180b11
Author: Rumesh Udash <allmesh67@gmail.com>
Date:   Tue Jul 23 16:57:38 2019 +0545

    Fix - Prevent empty search query

+++ includes/class-ur-form-handler.php
@@ -97,12 +97,18 @@ class UR_Form_Handler {
 		} elseif ( empty( $_POST['profile-pic-url'] ) ) {
 			$upload_dir  = wp_upload_dir();
 			$profile_url = get_user_meta( $user_id, 'user_registration_profile_pic_url', true );
-			$profile_url = $upload_dir['basedir'] . explode( '/uploads', $profile_url )[1];
 
-			if ( ! empty( $profile_url ) && file_exists( $profile_url ) ) {
-				@unlink( $profile_url );
+			// Check if profile already set?
+			if ( $profile_url ) {
+
+				// Then delete file and user meta.
+				$profile_url = $upload_dir['basedir'] . explode( '/uploads', $profile_url )[1];
+
+				if ( ! empty( $profile_url ) && file_exists( $profile_url ) ) {
+					@unlink( $profile_url );
+				}
+				delete_user_meta( $user_id, 'user_registration_profile_pic_url' );
 			}
-			delete_user_meta( $user_id, 'user_registration_profile_pic_url' );
 		}
 
 		$form_id_array = get_user_meta( $user_id, 'ur_form_id' );
@@ -202,7 +208,7 @@ class UR_Form_Handler {
 
 			ur_add_notice( __( 'User profile updated successfully.', 'user-registration' ) );
 
-			do_action( 'user_registration_save_profile_details', $user_id );
+			do_action( 'user_registration_save_profile_details', $user_id, $form_id );
 
 			wp_safe_redirect( ur_get_endpoint_url( 'edit-profile', '', ur_get_page_permalink( 'myaccount' ) ) );
 			exit;
